Graph for DWB model results

This function dynamically graphs the inputs and results of the DWBmodelUN.

Usage

graphDWB(var, tp, main = "", ...)

Arguments

var

It is a list that contains a time series of type ts which you want to graph.
For (tp = 2), it is recommended to list the simulated runoff series first, followed by the observed.
For (tp = 3), it must first contain the observed precipitation series, followed by the simulated runoff series and finally the observed runoff.
For (tp = 4), it must first contain the observed precipitation series, followed by the evapotranspiration series and finally the runoff time series.

tp

Variable which is defined to choose the type of graph.

main

Main title for the graph.

...

Other parameters of the dygraphs package.

Value

Prints a dynamic graph according to the requirements.

Details

It has three types of graphs:

(tp = 1): Plots any variable in a continuous format.

(tp = 2): Compares the runoff result of the model, with the observations.

(tp = 3): It allows to show a comparison between the observed and simulated runoff, as well as, with a dataset of precipitation.

(tp = 4): It presents a comparison between a set of precipitation, actual or potential evapotranspiration and runoff.

Author

Carolina Vega Viviescas <cvegav@unal.edu.co>
Pedro Felipe Arboleda Obando <pfarboledao@unal.edu.co>
David Zamora <dazamoraa@unal.edu.co>
Nicolas Duque Gardeazabal <nduqueg@unal.edu.co>

Water Resources Engineering Research Group - GIREH Universidad Nacional de Colombia - sede Bogota

Examples

# Example 1
data(P_sogamoso)
P.est <- ts(c(t(P_sogamoso[1, -2:-1])), start = c(2001, 1), frequency = 12)
var <- list("Precipitation" = P.est)

graphDWB(var, tp = 1, main = "Precipitation Lat:7.0 Lon:-72.94")

# Example 2
data(simDWB.sogamoso, EscSogObs)
runoff.sim <- ts(simDWB.sogamoso[,25], start = c(2001, 1), frequency = 12)
runoff.obs <- ts(EscSogObs[,25] , start = c(2001, 1), frequency = 12)
var <- list("Runoff.sim" = runoff.sim, "Runoff.obs" = runoff.obs)

graphDWB(var, tp = 2, main = "Runoff at basin closure: Gauge 24067010")

# Example 3
data(P_sogamoso, simDWB.sogamoso, EscSogObs)
P.est <- ts(c(t(P_sogamoso[1, 15:110])), start = c(2002, 1), frequency = 12)
runoff.sim <- ts(simDWB.sogamoso[13:108 ,25], start = c(2002, 1), frequency = 12)
runoff.obs <- ts(EscSogObs[13:108 ,25] , start = c(2002, 1), frequency = 12)
var <- list("Precipitation" = P.est,"Runoff.sim" = runoff.sim, "Runoff.obs" = runoff.obs)

graphDWB(var, tp = 3, main = "DWB results at Sogamoso Basin closure point")
#> <div id="htmlwidget-ac96cb3ee4656e2e9ec3" style="width:100%;height:150px;" class="dygraphs html-widget"></div>
#> <script type="application/json" data-for="htmlwidget-ac96cb3ee4656e2e9ec3">{"x":{"attrs":{"axes":{"x":{"pixelsPerLabel":60},"y":{"valueRange":[289,0]}},"series":{"Precipitation":{"axis":"y","strokeWidth":1.7}},"title":"DWB results at Sogamoso Basin closure point","labels":["month","Precipitation"],"retainDateWindow":false,"colors":["#2c7fb8","#2c7fb8"],"legend":"follow","labelsDivWidth":210,"labelsShowZeroValues":true,"labelsSeparateLines":false,"hideOverlayOnMouseOut":false,"ylabel":"P [mm/mth]"},"scale":"monthly","group":"A","annotations":[],"shadings":[],"events":[],"format":"date","data":[["2002-01-01T00:00:00.000Z","2002-02-01T00:00:00.000Z","2002-03-01T00:00:00.000Z","2002-04-01T00:00:00.000Z","2002-05-01T00:00:00.000Z","2002-06-01T00:00:00.000Z","2002-07-01T00:00:00.000Z","2002-08-01T00:00:00.000Z","2002-09-01T00:00:00.000Z","2002-10-01T00:00:00.000Z","2002-11-01T00:00:00.000Z","2002-12-01T00:00:00.000Z","2003-01-01T00:00:00.000Z","2003-02-01T00:00:00.000Z","2003-03-01T00:00:00.000Z","2003-04-01T00:00:00.000Z","2003-05-01T00:00:00.000Z","2003-06-01T00:00:00.000Z","2003-07-01T00:00:00.000Z","2003-08-01T00:00:00.000Z","2003-09-01T00:00:00.000Z","2003-10-01T00:00:00.000Z","2003-11-01T00:00:00.000Z","2003-12-01T00:00:00.000Z","2004-01-01T00:00:00.000Z","2004-02-01T00:00:00.000Z","2004-03-01T00:00:00.000Z","2004-04-01T00:00:00.000Z","2004-05-01T00:00:00.000Z","2004-06-01T00:00:00.000Z","2004-07-01T00:00:00.000Z","2004-08-01T00:00:00.000Z","2004-09-01T00:00:00.000Z","2004-10-01T00:00:00.000Z","2004-11-01T00:00:00.000Z","2004-12-01T00:00:00.000Z","2005-01-01T00:00:00.000Z","2005-02-01T00:00:00.000Z","2005-03-01T00:00:00.000Z","2005-04-01T00:00:00.000Z","2005-05-01T00:00:00.000Z","2005-06-01T00:00:00.000Z","2005-07-01T00:00:00.000Z","2005-08-01T00:00:00.000Z","2005-09-01T00:00:00.000Z","2005-10-01T00:00:00.000Z","2005-11-01T00:00:00.000Z","2005-12-01T00:00:00.000Z","2006-01-01T00:00:00.000Z","2006-02-01T00:00:00.000Z","2006-03-01T00:00:00.000Z","2006-04-01T00:00:00.000Z","2006-05-01T00:00:00.000Z","2006-06-01T00:00:00.000Z","2006-07-01T00:00:00.000Z","2006-08-01T00:00:00.000Z","2006-09-01T00:00:00.000Z","2006-10-01T00:00:00.000Z","2006-11-01T00:00:00.000Z","2006-12-01T00:00:00.000Z","2007-01-01T00:00:00.000Z","2007-02-01T00:00:00.000Z","2007-03-01T00:00:00.000Z","2007-04-01T00:00:00.000Z","2007-05-01T00:00:00.000Z","2007-06-01T00:00:00.000Z","2007-07-01T00:00:00.000Z","2007-08-01T00:00:00.000Z","2007-09-01T00:00:00.000Z","2007-10-01T00:00:00.000Z","2007-11-01T00:00:00.000Z","2007-12-01T00:00:00.000Z","2008-01-01T00:00:00.000Z","2008-02-01T00:00:00.000Z","2008-03-01T00:00:00.000Z","2008-04-01T00:00:00.000Z","2008-05-01T00:00:00.000Z","2008-06-01T00:00:00.000Z","2008-07-01T00:00:00.000Z","2008-08-01T00:00:00.000Z","2008-09-01T00:00:00.000Z","2008-10-01T00:00:00.000Z","2008-11-01T00:00:00.000Z","2008-12-01T00:00:00.000Z","2009-01-01T00:00:00.000Z","2009-02-01T00:00:00.000Z","2009-03-01T00:00:00.000Z","2009-04-01T00:00:00.000Z","2009-05-01T00:00:00.000Z","2009-06-01T00:00:00.000Z","2009-07-01T00:00:00.000Z","2009-08-01T00:00:00.000Z","2009-09-01T00:00:00.000Z","2009-10-01T00:00:00.000Z","2009-11-01T00:00:00.000Z","2009-12-01T00:00:00.000Z"],[21,21,105,213,137,170,99,62,106,104,53,46,33,77,100,221,94,200,186,96,167,187,103,73,62,44,66,167,141,82,116,74,178,156,158,58,84,100,9,143,187,197,146,115,122,187,173,34,58,44,126,239,221,149,142,117,115,164,106,29,52,18,133,150,139,118,135,189,121,198,73,30,63,77,133,129,187,114,180,142,129,148,158,18,72,61,173,123,107,153,106,89,50,164,130,24]],"plotter":"BarChart"},"evals":[],"jsHooks":[]}</script>
#> <div id="htmlwidget-e5c8c404fe174e4c81bd" style="width:100%;height:300px;" class="dygraphs html-widget"></div>
#> <script type="application/json" data-for="htmlwidget-e5c8c404fe174e4c81bd">{"x":{"attrs":{"ylabel":"Runoff [mm/mth]","labels":["month","Runoff.sim","Runoff.obs"],"retainDateWindow":false,"axes":{"x":{"pixelsPerLabel":60}},"colors":["#ef8a62","#404040"],"series":{"Runoff.sim":{"axis":"y","strokeWidth":1.7},"Runoff.obs":{"axis":"y","drawPoints":true,"pointSize":2,"strokeWidth":1.7}},"legend":"follow","labelsDivWidth":210,"labelsShowZeroValues":true,"labelsSeparateLines":false,"highlightCircleSize":3,"highlightSeriesBackgroundAlpha":0.2,"highlightSeriesOpts":[],"hideOverlayOnMouseOut":false,"showRangeSelector":true,"rangeSelectorHeight":25,"rangeSelectorPlotFillColor":" #A7B1C4","rangeSelectorPlotStrokeColor":"#808FAB","interactionModel":"Dygraph.Interaction.defaultModel"},"scale":"monthly","group":"A","annotations":[],"shadings":[],"events":[],"format":"date","data":[["2002-01-01T00:00:00.000Z","2002-02-01T00:00:00.000Z","2002-03-01T00:00:00.000Z","2002-04-01T00:00:00.000Z","2002-05-01T00:00:00.000Z","2002-06-01T00:00:00.000Z","2002-07-01T00:00:00.000Z","2002-08-01T00:00:00.000Z","2002-09-01T00:00:00.000Z","2002-10-01T00:00:00.000Z","2002-11-01T00:00:00.000Z","2002-12-01T00:00:00.000Z","2003-01-01T00:00:00.000Z","2003-02-01T00:00:00.000Z","2003-03-01T00:00:00.000Z","2003-04-01T00:00:00.000Z","2003-05-01T00:00:00.000Z","2003-06-01T00:00:00.000Z","2003-07-01T00:00:00.000Z","2003-08-01T00:00:00.000Z","2003-09-01T00:00:00.000Z","2003-10-01T00:00:00.000Z","2003-11-01T00:00:00.000Z","2003-12-01T00:00:00.000Z","2004-01-01T00:00:00.000Z","2004-02-01T00:00:00.000Z","2004-03-01T00:00:00.000Z","2004-04-01T00:00:00.000Z","2004-05-01T00:00:00.000Z","2004-06-01T00:00:00.000Z","2004-07-01T00:00:00.000Z","2004-08-01T00:00:00.000Z","2004-09-01T00:00:00.000Z","2004-10-01T00:00:00.000Z","2004-11-01T00:00:00.000Z","2004-12-01T00:00:00.000Z","2005-01-01T00:00:00.000Z","2005-02-01T00:00:00.000Z","2005-03-01T00:00:00.000Z","2005-04-01T00:00:00.000Z","2005-05-01T00:00:00.000Z","2005-06-01T00:00:00.000Z","2005-07-01T00:00:00.000Z","2005-08-01T00:00:00.000Z","2005-09-01T00:00:00.000Z","2005-10-01T00:00:00.000Z","2005-11-01T00:00:00.000Z","2005-12-01T00:00:00.000Z","2006-01-01T00:00:00.000Z","2006-02-01T00:00:00.000Z","2006-03-01T00:00:00.000Z","2006-04-01T00:00:00.000Z","2006-05-01T00:00:00.000Z","2006-06-01T00:00:00.000Z","2006-07-01T00:00:00.000Z","2006-08-01T00:00:00.000Z","2006-09-01T00:00:00.000Z","2006-10-01T00:00:00.000Z","2006-11-01T00:00:00.000Z","2006-12-01T00:00:00.000Z","2007-01-01T00:00:00.000Z","2007-02-01T00:00:00.000Z","2007-03-01T00:00:00.000Z","2007-04-01T00:00:00.000Z","2007-05-01T00:00:00.000Z","2007-06-01T00:00:00.000Z","2007-07-01T00:00:00.000Z","2007-08-01T00:00:00.000Z","2007-09-01T00:00:00.000Z","2007-10-01T00:00:00.000Z","2007-11-01T00:00:00.000Z","2007-12-01T00:00:00.000Z","2008-01-01T00:00:00.000Z","2008-02-01T00:00:00.000Z","2008-03-01T00:00:00.000Z","2008-04-01T00:00:00.000Z","2008-05-01T00:00:00.000Z","2008-06-01T00:00:00.000Z","2008-07-01T00:00:00.000Z","2008-08-01T00:00:00.000Z","2008-09-01T00:00:00.000Z","2008-10-01T00:00:00.000Z","2008-11-01T00:00:00.000Z","2008-12-01T00:00:00.000Z","2009-01-01T00:00:00.000Z","2009-02-01T00:00:00.000Z","2009-03-01T00:00:00.000Z","2009-04-01T00:00:00.000Z","2009-05-01T00:00:00.000Z","2009-06-01T00:00:00.000Z","2009-07-01T00:00:00.000Z","2009-08-01T00:00:00.000Z","2009-09-01T00:00:00.000Z","2009-10-01T00:00:00.000Z","2009-11-01T00:00:00.000Z","2009-12-01T00:00:00.000Z"],[45.20216725721488,37.05566649015511,70.55376786670575,99.32124389232642,92.81944189784932,76.74157382721893,59.52440917967716,56.98786191453079,61.90548328833464,71.76541515147207,64.78631406319334,54.51459314052763,36.99997452794864,38.23159068536734,51.5603121475018,91.07235190395512,87.87798419596626,79.30126239487926,66.71907391384511,69.63631299995289,78.07840486675858,104.8894580996532,101.4740718936371,66.58653175671402,47.67602613199988,40.52946798251484,45.47341550947447,93.94159029708223,112.5196089206925,69.11240800443569,60.92689543929973,51.56085464865838,74.71164068177075,112.7364126462114,96.84174097930268,60.02396445108742,54.88129013020743,51.24001729586295,45.67512085901091,78.30653415008935,98.86902499895284,79.25692454779765,67.2742894360772,60.27511778469578,62.56539360517516,107.4546515034035,113.5960303776137,69.16275192531479,54.63024975641794,40.6609816421734,81.79214011650831,115.6649377745107,123.1472008159262,94.66708224571531,67.61755851141885,60.56244434986514,62.52125164034728,100.2042534214122,103.9358454062908,76.86173279064455,54.19283391676794,37.91735998798998,68.65856391531702,91.41121458381139,102.3442617140272,73.26033390881666,70.19376894663728,82.00675846476776,71.02395702555306,129.9471804031851,88.75492558647262,74.47024925712832,53.81955116282793,49.34435375802057,54.69109958282029,70.42347735473959,115.2294823751653,86.31590967065067,72.27860862541213,92.7647425422272,87.31039294201442,104.903001873375,113.895791340937,65.05867545506005,55.21792678211264,58.00482183643164,97.24596131492599,88.27706368636733,79.200091116292,73.94396011924923,55.7909946956004,65.39981633116683,52.49959305372516,71.7000645611604,69.99034961509159,50.6970345989148],[14.54811562802056,12.5835422025244,40.88367180901805,82.62764701351482,68.06669103630801,84.55369939145221,33.46194997936591,29.69972766779484,33.5903534712284,42.37315231462298,48.16414979762146,33.60319382041465,10.48671318040988,13.54656839149311,19.44028866798158,70.4806766833229,65.65270538929312,67.79704370339677,40.28017539726433,45.58323961118533,53.26176842456246,84.45097659796224,82.22959618874108,50.4368916035876,23.97293193072761,12.90455093218064,17.50139594085792,57.51192400521101,86.24862548403716,51.69524582384004,33.98840429600213,32.80709217086719,50.79642138080258,79.59732460555996,101.5800024124189,38.44400546363067,25.69353872168502,33.5903534712284,18.70838876436536,48.36959538460145,74.93627785095144,53.06916318676873,36.42807064138952,33.24366404319967,36.6848776251145,77.8381967670438,97.99754498945535,47.81746036959272,30.59855211083229,18.28465724121913,48.24119189273895,83.00001713991604,106.3309316113312,76.18179172201762,37.6735845124557,32.49892379039721,40.15177190540184,74.28142004245271,90.44741966794068,58.82163962220844,22.76593910722017,10.31593653623276,35.51640584916581,75.36000937409766,77.92807921134754,57.29363806904477,43.1307329166117,53.31312982130746,51.07890906290007,103.4290126952388,78.40317213123878,47.41940954481899,28.46705414591489,20.99397091951775,24.70483183434382,39.59963689039311,82.29379793467233,61.69787783992831,49.14001633577641,60.16987628676463,67.48887532292679,85.1443554540197,91.14079852399816,42.38599266380924,34.84870769148083,43.91399421697291,76.96505302237883,73.83200782093398,72.18844312509407,58.15394146452348,34.03976569274712,48.87036900286517,36.67203727592825,45.2879115799016,70.33943284227415,24.93595811969631]]},"evals":["attrs.interactionModel"],"jsHooks":[]}</script>

# Example 4
data(P_sogamoso, PET_sogamoso, simDWB.sogamoso)
P <- ts(c(t(P_sogamoso[1, -2:-1])), start = c(2001, 1), frequency = 12)
PET <- ts(c(t(PET_sogamoso[1, -2:-1])), start = c(2001, 1), frequency = 12)
runoff.sim <- ts(simDWB.sogamoso[ ,25], start = c(2001, 1), frequency = 12)
var <- list("P" = P,"PET" = PET, "Runoff.sim" = runoff.sim)

graphDWB(var, tp = 4, main = "General Comparison Sogamoso Basin")
#> <div id="htmlwidget-36aa3d2a04d42bbc2145" style="width:100%;height:160px;" class="dygraphs html-widget"></div>
#> <script type="application/json" data-for="htmlwidget-36aa3d2a04d42bbc2145">{"x":{"attrs":{"axes":{"x":{"pixelsPerLabel":60},"y":{"valueRange":[305,0]}},"series":{"P":{"axis":"y"}},"title":"General Comparison Sogamoso Basin","labels":["month","P"],"retainDateWindow":false,"colors":["#2c7fb8","#2c7fb8"],"legend":"follow","labelsDivWidth":210,"labelsShowZeroValues":true,"labelsSeparateLines":false,"hideOverlayOnMouseOut":false,"ylabel":"P [mm/mth]"},"scale":"monthly","group":"A","annotations":[],"shadings":[],"events":[],"format":"date","data":[["2001-01-01T00:00:00.000Z","2001-02-01T00:00:00.000Z","2001-03-01T00:00:00.000Z","2001-04-01T00:00:00.000Z","2001-05-01T00:00:00.000Z","2001-06-01T00:00:00.000Z","2001-07-01T00:00:00.000Z","2001-08-01T00:00:00.000Z","2001-09-01T00:00:00.000Z","2001-10-01T00:00:00.000Z","2001-11-01T00:00:00.000Z","2001-12-01T00:00:00.000Z","2002-01-01T00:00:00.000Z","2002-02-01T00:00:00.000Z","2002-03-01T00:00:00.000Z","2002-04-01T00:00:00.000Z","2002-05-01T00:00:00.000Z","2002-06-01T00:00:00.000Z","2002-07-01T00:00:00.000Z","2002-08-01T00:00:00.000Z","2002-09-01T00:00:00.000Z","2002-10-01T00:00:00.000Z","2002-11-01T00:00:00.000Z","2002-12-01T00:00:00.000Z","2003-01-01T00:00:00.000Z","2003-02-01T00:00:00.000Z","2003-03-01T00:00:00.000Z","2003-04-01T00:00:00.000Z","2003-05-01T00:00:00.000Z","2003-06-01T00:00:00.000Z","2003-07-01T00:00:00.000Z","2003-08-01T00:00:00.000Z","2003-09-01T00:00:00.000Z","2003-10-01T00:00:00.000Z","2003-11-01T00:00:00.000Z","2003-12-01T00:00:00.000Z","2004-01-01T00:00:00.000Z","2004-02-01T00:00:00.000Z","2004-03-01T00:00:00.000Z","2004-04-01T00:00:00.000Z","2004-05-01T00:00:00.000Z","2004-06-01T00:00:00.000Z","2004-07-01T00:00:00.000Z","2004-08-01T00:00:00.000Z","2004-09-01T00:00:00.000Z","2004-10-01T00:00:00.000Z","2004-11-01T00:00:00.000Z","2004-12-01T00:00:00.000Z","2005-01-01T00:00:00.000Z","2005-02-01T00:00:00.000Z","2005-03-01T00:00:00.000Z","2005-04-01T00:00:00.000Z","2005-05-01T00:00:00.000Z","2005-06-01T00:00:00.000Z","2005-07-01T00:00:00.000Z","2005-08-01T00:00:00.000Z","2005-09-01T00:00:00.000Z","2005-10-01T00:00:00.000Z","2005-11-01T00:00:00.000Z","2005-12-01T00:00:00.000Z","2006-01-01T00:00:00.000Z","2006-02-01T00:00:00.000Z","2006-03-01T00:00:00.000Z","2006-04-01T00:00:00.000Z","2006-05-01T00:00:00.000Z","2006-06-01T00:00:00.000Z","2006-07-01T00:00:00.000Z","2006-08-01T00:00:00.000Z","2006-09-01T00:00:00.000Z","2006-10-01T00:00:00.000Z","2006-11-01T00:00:00.000Z","2006-12-01T00:00:00.000Z","2007-01-01T00:00:00.000Z","2007-02-01T00:00:00.000Z","2007-03-01T00:00:00.000Z","2007-04-01T00:00:00.000Z","2007-05-01T00:00:00.000Z","2007-06-01T00:00:00.000Z","2007-07-01T00:00:00.000Z","2007-08-01T00:00:00.000Z","2007-09-01T00:00:00.000Z","2007-10-01T00:00:00.000Z","2007-11-01T00:00:00.000Z","2007-12-01T00:00:00.000Z","2008-01-01T00:00:00.000Z","2008-02-01T00:00:00.000Z","2008-03-01T00:00:00.000Z","2008-04-01T00:00:00.000Z","2008-05-01T00:00:00.000Z","2008-06-01T00:00:00.000Z","2008-07-01T00:00:00.000Z","2008-08-01T00:00:00.000Z","2008-09-01T00:00:00.000Z","2008-10-01T00:00:00.000Z","2008-11-01T00:00:00.000Z","2008-12-01T00:00:00.000Z","2009-01-01T00:00:00.000Z","2009-02-01T00:00:00.000Z","2009-03-01T00:00:00.000Z","2009-04-01T00:00:00.000Z","2009-05-01T00:00:00.000Z","2009-06-01T00:00:00.000Z","2009-07-01T00:00:00.000Z","2009-08-01T00:00:00.000Z","2009-09-01T00:00:00.000Z","2009-10-01T00:00:00.000Z","2009-11-01T00:00:00.000Z","2009-12-01T00:00:00.000Z","2010-01-01T00:00:00.000Z","2010-02-01T00:00:00.000Z","2010-03-01T00:00:00.000Z","2010-04-01T00:00:00.000Z","2010-05-01T00:00:00.000Z","2010-06-01T00:00:00.000Z","2010-07-01T00:00:00.000Z","2010-08-01T00:00:00.000Z","2010-09-01T00:00:00.000Z","2010-10-01T00:00:00.000Z","2010-11-01T00:00:00.000Z","2010-12-01T00:00:00.000Z"],[28,42,75,78,105,94,116,68,107,152,113,65,21,21,105,213,137,170,99,62,106,104,53,46,33,77,100,221,94,200,186,96,167,187,103,73,62,44,66,167,141,82,116,74,178,156,158,58,84,100,9,143,187,197,146,115,122,187,173,34,58,44,126,239,221,149,142,117,115,164,106,29,52,18,133,150,139,118,135,189,121,198,73,30,63,77,133,129,187,114,180,142,129,148,158,18,72,61,173,123,107,153,106,89,50,164,130,24,12,27,37,161,221,212,244,152,252,169,255,109]],"plotter":"BarChart"},"evals":[],"jsHooks":[]}</script>
#> <div id="htmlwidget-febe03efa1a2d8d52a86" style="width:100%;height:140px;" class="dygraphs html-widget"></div>
#> <script type="application/json" data-for="htmlwidget-febe03efa1a2d8d52a86">{"x":{"attrs":{"axes":{"x":{"pixelsPerLabel":60},"y":{"valueRange":[0,147.3]}},"series":{"PET":{"axis":"y","strokeWidth":1.7}},"labels":["month","PET"],"retainDateWindow":false,"colors":["#1a9850","#1a9850"],"legend":"follow","labelsDivWidth":210,"labelsShowZeroValues":true,"labelsSeparateLines":false,"hideOverlayOnMouseOut":false,"ylabel":"ET [mm/mth]"},"scale":"monthly","group":"A","annotations":[],"shadings":[],"events":[],"format":"date","data":[["2001-01-01T00:00:00.000Z","2001-02-01T00:00:00.000Z","2001-03-01T00:00:00.000Z","2001-04-01T00:00:00.000Z","2001-05-01T00:00:00.000Z","2001-06-01T00:00:00.000Z","2001-07-01T00:00:00.000Z","2001-08-01T00:00:00.000Z","2001-09-01T00:00:00.000Z","2001-10-01T00:00:00.000Z","2001-11-01T00:00:00.000Z","2001-12-01T00:00:00.000Z","2002-01-01T00:00:00.000Z","2002-02-01T00:00:00.000Z","2002-03-01T00:00:00.000Z","2002-04-01T00:00:00.000Z","2002-05-01T00:00:00.000Z","2002-06-01T00:00:00.000Z","2002-07-01T00:00:00.000Z","2002-08-01T00:00:00.000Z","2002-09-01T00:00:00.000Z","2002-10-01T00:00:00.000Z","2002-11-01T00:00:00.000Z","2002-12-01T00:00:00.000Z","2003-01-01T00:00:00.000Z","2003-02-01T00:00:00.000Z","2003-03-01T00:00:00.000Z","2003-04-01T00:00:00.000Z","2003-05-01T00:00:00.000Z","2003-06-01T00:00:00.000Z","2003-07-01T00:00:00.000Z","2003-08-01T00:00:00.000Z","2003-09-01T00:00:00.000Z","2003-10-01T00:00:00.000Z","2003-11-01T00:00:00.000Z","2003-12-01T00:00:00.000Z","2004-01-01T00:00:00.000Z","2004-02-01T00:00:00.000Z","2004-03-01T00:00:00.000Z","2004-04-01T00:00:00.000Z","2004-05-01T00:00:00.000Z","2004-06-01T00:00:00.000Z","2004-07-01T00:00:00.000Z","2004-08-01T00:00:00.000Z","2004-09-01T00:00:00.000Z","2004-10-01T00:00:00.000Z","2004-11-01T00:00:00.000Z","2004-12-01T00:00:00.000Z","2005-01-01T00:00:00.000Z","2005-02-01T00:00:00.000Z","2005-03-01T00:00:00.000Z","2005-04-01T00:00:00.000Z","2005-05-01T00:00:00.000Z","2005-06-01T00:00:00.000Z","2005-07-01T00:00:00.000Z","2005-08-01T00:00:00.000Z","2005-09-01T00:00:00.000Z","2005-10-01T00:00:00.000Z","2005-11-01T00:00:00.000Z","2005-12-01T00:00:00.000Z","2006-01-01T00:00:00.000Z","2006-02-01T00:00:00.000Z","2006-03-01T00:00:00.000Z","2006-04-01T00:00:00.000Z","2006-05-01T00:00:00.000Z","2006-06-01T00:00:00.000Z","2006-07-01T00:00:00.000Z","2006-08-01T00:00:00.000Z","2006-09-01T00:00:00.000Z","2006-10-01T00:00:00.000Z","2006-11-01T00:00:00.000Z","2006-12-01T00:00:00.000Z","2007-01-01T00:00:00.000Z","2007-02-01T00:00:00.000Z","2007-03-01T00:00:00.000Z","2007-04-01T00:00:00.000Z","2007-05-01T00:00:00.000Z","2007-06-01T00:00:00.000Z","2007-07-01T00:00:00.000Z","2007-08-01T00:00:00.000Z","2007-09-01T00:00:00.000Z","2007-10-01T00:00:00.000Z","2007-11-01T00:00:00.000Z","2007-12-01T00:00:00.000Z","2008-01-01T00:00:00.000Z","2008-02-01T00:00:00.000Z","2008-03-01T00:00:00.000Z","2008-04-01T00:00:00.000Z","2008-05-01T00:00:00.000Z","2008-06-01T00:00:00.000Z","2008-07-01T00:00:00.000Z","2008-08-01T00:00:00.000Z","2008-09-01T00:00:00.000Z","2008-10-01T00:00:00.000Z","2008-11-01T00:00:00.000Z","2008-12-01T00:00:00.000Z","2009-01-01T00:00:00.000Z","2009-02-01T00:00:00.000Z","2009-03-01T00:00:00.000Z","2009-04-01T00:00:00.000Z","2009-05-01T00:00:00.000Z","2009-06-01T00:00:00.000Z","2009-07-01T00:00:00.000Z","2009-08-01T00:00:00.000Z","2009-09-01T00:00:00.000Z","2009-10-01T00:00:00.000Z","2009-11-01T00:00:00.000Z","2009-12-01T00:00:00.000Z","2010-01-01T00:00:00.000Z","2010-02-01T00:00:00.000Z","2010-03-01T00:00:00.000Z","2010-04-01T00:00:00.000Z","2010-05-01T00:00:00.000Z","2010-06-01T00:00:00.000Z","2010-07-01T00:00:00.000Z","2010-08-01T00:00:00.000Z","2010-09-01T00:00:00.000Z","2010-10-01T00:00:00.000Z","2010-11-01T00:00:00.000Z","2010-12-01T00:00:00.000Z"],[81.92,82.97,85.40000000000001,86.58,81.94,75.59,79.90000000000001,79.13,85.77,80.15000000000001,75.79000000000001,74.25,82.48999999999999,89.12,87.52,79.18000000000001,80.53,76.7,78.01000000000001,83.73999999999999,87.22,82.75,75.78,77.66,86.97,87.53,88.13,82.73,83.09,80.23,78.09999999999999,84.95,85.58,79.61,76.34999999999999,74.15000000000001,79.33,92.03,89.45999999999999,82.02,80.18000000000001,78.62,80.86,82.47,85.15000000000001,80.56,74.34999999999999,74.58,76.68000000000001,85.62,97.3,86.52,83.73,78.84999999999999,82.36,81.66,84.56999999999999,80.95999999999999,74.48,76.45999999999999,78.76000000000001,91.5,82.70999999999999,80.48,81.39,78.08,74.88,84.53,85.47,78.34999999999999,74.27,73.45999999999999,81.55,93.73999999999999,86.09,88.3,84.61,75.91,82.26000000000001,79.31,83.59,78.95,76.94,72.31,77.59,80.75,85.56,82.88,78.54000000000001,79.69,75.12,79.98999999999999,83.23999999999999,80.05,71.5,74.62,75.38,78.12,81.28,82.97,82.34,79.8,76.44,83.48,87.68000000000001,82.58,77.84,78.90000000000001,87.84999999999999,90.72,93.78,86.18000000000001,82.62,78.62,75.87,80.78,79.55,77.48,69.59,70.20999999999999]],"plotter":"BarChart"},"evals":[],"jsHooks":[]}</script>
#> <div id="htmlwidget-1fb4450895fe099f74a1" style="width:100%;height:225px;" class="dygraphs html-widget"></div>
#> <script type="application/json" data-for="htmlwidget-1fb4450895fe099f74a1">{"x":{"attrs":{"ylabel":"Runoff [mm/mth]","labels":["month","Runoff.sim"],"retainDateWindow":false,"axes":{"x":{"pixelsPerLabel":60}},"colors":["#ef8a62","#ef8a62"],"series":{"Runoff.sim":{"axis":"y","strokeWidth":1.7}},"legend":"follow","labelsDivWidth":210,"labelsShowZeroValues":true,"labelsSeparateLines":false,"hideOverlayOnMouseOut":false,"showRangeSelector":true,"rangeSelectorHeight":25,"rangeSelectorPlotFillColor":" #A7B1C4","rangeSelectorPlotStrokeColor":"#808FAB","interactionModel":"Dygraph.Interaction.defaultModel"},"scale":"monthly","group":"A","annotations":[],"shadings":[],"events":[],"format":"date","data":[["2001-01-01T00:00:00.000Z","2001-02-01T00:00:00.000Z","2001-03-01T00:00:00.000Z","2001-04-01T00:00:00.000Z","2001-05-01T00:00:00.000Z","2001-06-01T00:00:00.000Z","2001-07-01T00:00:00.000Z","2001-08-01T00:00:00.000Z","2001-09-01T00:00:00.000Z","2001-10-01T00:00:00.000Z","2001-11-01T00:00:00.000Z","2001-12-01T00:00:00.000Z","2002-01-01T00:00:00.000Z","2002-02-01T00:00:00.000Z","2002-03-01T00:00:00.000Z","2002-04-01T00:00:00.000Z","2002-05-01T00:00:00.000Z","2002-06-01T00:00:00.000Z","2002-07-01T00:00:00.000Z","2002-08-01T00:00:00.000Z","2002-09-01T00:00:00.000Z","2002-10-01T00:00:00.000Z","2002-11-01T00:00:00.000Z","2002-12-01T00:00:00.000Z","2003-01-01T00:00:00.000Z","2003-02-01T00:00:00.000Z","2003-03-01T00:00:00.000Z","2003-04-01T00:00:00.000Z","2003-05-01T00:00:00.000Z","2003-06-01T00:00:00.000Z","2003-07-01T00:00:00.000Z","2003-08-01T00:00:00.000Z","2003-09-01T00:00:00.000Z","2003-10-01T00:00:00.000Z","2003-11-01T00:00:00.000Z","2003-12-01T00:00:00.000Z","2004-01-01T00:00:00.000Z","2004-02-01T00:00:00.000Z","2004-03-01T00:00:00.000Z","2004-04-01T00:00:00.000Z","2004-05-01T00:00:00.000Z","2004-06-01T00:00:00.000Z","2004-07-01T00:00:00.000Z","2004-08-01T00:00:00.000Z","2004-09-01T00:00:00.000Z","2004-10-01T00:00:00.000Z","2004-11-01T00:00:00.000Z","2004-12-01T00:00:00.000Z","2005-01-01T00:00:00.000Z","2005-02-01T00:00:00.000Z","2005-03-01T00:00:00.000Z","2005-04-01T00:00:00.000Z","2005-05-01T00:00:00.000Z","2005-06-01T00:00:00.000Z","2005-07-01T00:00:00.000Z","2005-08-01T00:00:00.000Z","2005-09-01T00:00:00.000Z","2005-10-01T00:00:00.000Z","2005-11-01T00:00:00.000Z","2005-12-01T00:00:00.000Z","2006-01-01T00:00:00.000Z","2006-02-01T00:00:00.000Z","2006-03-01T00:00:00.000Z","2006-04-01T00:00:00.000Z","2006-05-01T00:00:00.000Z","2006-06-01T00:00:00.000Z","2006-07-01T00:00:00.000Z","2006-08-01T00:00:00.000Z","2006-09-01T00:00:00.000Z","2006-10-01T00:00:00.000Z","2006-11-01T00:00:00.000Z","2006-12-01T00:00:00.000Z","2007-01-01T00:00:00.000Z","2007-02-01T00:00:00.000Z","2007-03-01T00:00:00.000Z","2007-04-01T00:00:00.000Z","2007-05-01T00:00:00.000Z","2007-06-01T00:00:00.000Z","2007-07-01T00:00:00.000Z","2007-08-01T00:00:00.000Z","2007-09-01T00:00:00.000Z","2007-10-01T00:00:00.000Z","2007-11-01T00:00:00.000Z","2007-12-01T00:00:00.000Z","2008-01-01T00:00:00.000Z","2008-02-01T00:00:00.000Z","2008-03-01T00:00:00.000Z","2008-04-01T00:00:00.000Z","2008-05-01T00:00:00.000Z","2008-06-01T00:00:00.000Z","2008-07-01T00:00:00.000Z","2008-08-01T00:00:00.000Z","2008-09-01T00:00:00.000Z","2008-10-01T00:00:00.000Z","2008-11-01T00:00:00.000Z","2008-12-01T00:00:00.000Z","2009-01-01T00:00:00.000Z","2009-02-01T00:00:00.000Z","2009-03-01T00:00:00.000Z","2009-04-01T00:00:00.000Z","2009-05-01T00:00:00.000Z","2009-06-01T00:00:00.000Z","2009-07-01T00:00:00.000Z","2009-08-01T00:00:00.000Z","2009-09-01T00:00:00.000Z","2009-10-01T00:00:00.000Z","2009-11-01T00:00:00.000Z","2009-12-01T00:00:00.000Z","2010-01-01T00:00:00.000Z","2010-02-01T00:00:00.000Z","2010-03-01T00:00:00.000Z","2010-04-01T00:00:00.000Z","2010-05-01T00:00:00.000Z","2010-06-01T00:00:00.000Z","2010-07-01T00:00:00.000Z","2010-08-01T00:00:00.000Z","2010-09-01T00:00:00.000Z","2010-10-01T00:00:00.000Z","2010-11-01T00:00:00.000Z","2010-12-01T00:00:00.000Z"],[241.9102912730664,89.73073979127373,76.58711330305155,67.51734446237901,80.48898878946433,68.4826347865601,62.94353445652727,57.88827769748384,75.05714218069026,95.17083310744243,85.36387298166284,70.89918237318732,45.20216725721488,37.05566649015511,70.55376786670575,99.32124389232642,92.81944189784932,76.74157382721893,59.52440917967716,56.98786191453079,61.90548328833464,71.76541515147207,64.78631406319334,54.51459314052763,36.99997452794864,38.23159068536734,51.5603121475018,91.07235190395512,87.87798419596626,79.30126239487926,66.71907391384511,69.63631299995289,78.07840486675858,104.8894580996532,101.4740718936371,66.58653175671402,47.67602613199988,40.52946798251484,45.47341550947447,93.94159029708223,112.5196089206925,69.11240800443569,60.92689543929973,51.56085464865838,74.71164068177075,112.7364126462114,96.84174097930268,60.02396445108742,54.88129013020743,51.24001729586295,45.67512085901091,78.30653415008935,98.86902499895284,79.25692454779765,67.2742894360772,60.27511778469578,62.56539360517516,107.4546515034035,113.5960303776137,69.16275192531479,54.63024975641794,40.6609816421734,81.79214011650831,115.6649377745107,123.1472008159262,94.66708224571531,67.61755851141885,60.56244434986514,62.52125164034728,100.2042534214122,103.9358454062908,76.86173279064455,54.19283391676794,37.91735998798998,68.65856391531702,91.41121458381139,102.3442617140272,73.26033390881666,70.19376894663728,82.00675846476776,71.02395702555306,129.9471804031851,88.75492558647262,74.47024925712832,53.81955116282793,49.34435375802057,54.69109958282029,70.42347735473959,115.2294823751653,86.31590967065067,72.27860862541213,92.7647425422272,87.31039294201442,104.903001873375,113.895791340937,65.05867545506005,55.21792678211264,58.00482183643164,97.24596131492599,88.27706368636733,79.200091116292,73.94396011924923,55.7909946956004,65.39981633116683,52.49959305372516,71.7000645611604,69.99034961509159,50.6970345989148,35.04810844819482,33.88042567092781,46.03756482703834,89.28819597482247,114.3044024020135,99.50095789712893,136.908478097222,105.6050983030391,128.7486255280824,129.7140013607388,153.2657619256734,94.861938371223]]},"evals":["attrs.interactionModel"],"jsHooks":[]}</script>